pasta-do-tita/projetos/video-factory/requirements.txt
moviepy>=2.0
pillow
requests
openai-whisper
piper-tts
edge-tts
python-dotenv
